feat: déduplication de code + fix test sur adresse existante
